backend/requirements.txt
fastapi
uvicorn
langchain
langchain-anthropic
python-dotenv